bug fix with logout, design buttons, refact displayminership->color,name etc
