+++ frontend/src/components/station/Station.css
@@ -2,7 +2,7 @@
     display: flex;
     flex-flow: row;
     font-size: 18px;
-    width: 70vw;
+    width: 80vw;
     margin: auto;
     min-width: 1200px;
-}
+    }
