+++ frontend/src/components/station/Station.jsx
@@ -1,5 +1,5 @@
-import StationBar from './station/StationBar';
-import Messages from "./station/Messages";
+import StationBar from './stationbar/StationBar';
+import Messages from "./messages/Messages";
 import "./Station.css";
 import { StationContext } from "./MessageContext";
 import { StorageStationContext } from "./StorageContext";
@@ -7,7 +7,9 @@ import { HangarStationContext } from "./HangarContext";
 import { useOutletContext } from 'react-router-dom';
 
 const Station = () => {
-    const [, , user, , stationId] = useOutletContext()
+    const context = useOutletContext()
+    const user = context.user;
+    const stationId = context.stationId;
 
     if (user === null || stationId === null) {
         return <div>Loading...</div>
