+++ frontend/src/components/station/messages/DisplayMinerShip.css
@@ -7,9 +7,7 @@
     justify-items: center;
 }
 
-.ship-name {
-    height: 65px !important;
-}
+
 
 .container {
     width: 50%;
