+++ frontend/src/components/station/messages/ShipColor.css
@@ -1,0 +1,5 @@
+.ship-color {
+    display: flex;
+    flex-flow: row;
+    justify-content: space-between;
+}
