+++ frontend/src/components/station/stationbar/Hangar.css
@@ -5,7 +5,6 @@
     padding: 5px;
     border-radius: 5px;
     color: white;
-    font-family: Garamond, serif;
     font-size: 18px;
     border: #000000 solid 4px;
     box-shadow:  0 0 5px 5px rgba(0, 23, 0, 0.85);
@@ -17,5 +16,10 @@
     justify-content: flex-start;
     align-items: center;
     justify-items: center;
+}
+
+.shiplist:hover {
+    text-shadow: 0px 0px .3em rgba(236, 200, 90, 0.85), 0px 0px .3em rgb(225, 196, 132), 0px 0px .3em rgb(225, 196, 132);
+    cursor: pointer;
 }
