+++ frontend/src/components/station/stationbar/StationBar.css
@@ -8,19 +8,6 @@
     height: 80vh;
 }
 
-.button {
-    margin: 0;
-    padding: 4px;
-    font-size: 15px;
-    border: black solid 2px;
-    border-radius: 5px;
-    background-color: forestgreen;
-    width: 75px;
-    text-align: center;
-    justify-self: center;
-    align-self: center;
-}
-
 .menu {
     display: flex;
     flex-flow: row;
