+++ frontend/src/components/station/stationbar/Storage.css
@@ -5,7 +5,6 @@
     padding: 5px;
     border-radius: 5px;
     color: white;
-    font-family: Garamond, serif;
     font-size: 18px;
     border: #000000 solid 4px;
     box-shadow: 0 0 5px 5px rgba(0, 23, 0, 0.85);
