+++ frontend/src/index.css
@@ -1,14 +1,50 @@
-body {
+@import url('https://fonts.googleapis.com/css2?family=Orbitron&display=swap');
 
+body {
     margin: 0;
     padding: 0;
     height: 100vh;
     background-image: url("./../public/sanc-bridge-full-transparent.jpg");
     /*background-color: forestgreen;*/
+    font-family: 'Orbitron', sans-serif;
 }
 
 #root {
     height: 100vh;
 }
 
+.clickable {
+    text-decoration: underline;
+    color: #e0b861;
+}
 
+.clickable:hover {
+    text-shadow: 1px 1px 1px rgb(24, 17, 6), 0 0 .2em rgb(14, 12, 1),
+    0 0 0.2em rgb(44, 33, 5);
+    cursor: pointer;
+}
+
+h2 {
+    margin:   0 0 15px 0;
+}
+
+.button {
+    margin: 0 5px;
+    padding: 4px;
+    font-size: 15px;
+    border: black solid 2px;
+    border-radius: 5px;
+    background-color: forestgreen;
+    width: 75px;
+    text-align: center;
+    justify-self: center;
+    align-self: center;
+    box-shadow: 1px 1px 1px rgb(21, 49, 1), 0px 0px .2em rgb(21, 49, 1), 0px 0px .2em rgb(21, 49, 1);
+}
+
+.button:hover {
+    box-shadow:  0px 0px .3em rgba(236, 200, 90, 0.85), 0px 0px .3em rgb(225, 196, 132), 0px 0px .3em rgb(225, 196, 132);
+    border: forestgreen solid 2px;
+    background-color: #2db62d;
+    cursor: pointer;
+}
